+++ scripts/archiso-base-211216.sh
@@ -279,20 +279,6 @@ Depends = sed
 Exec = /usr/bin/sed -i "s/#Server/Server/g" /etc/pacman.d/mirrorlist
 _
 
-teew etc/pacman.d/hooks/update-keyring.hook << '_'
-# remove from airootfs!
-[Trigger]
-Operation = Install
-Type = Package
-Target = pacman
-
-[Action]
-Description = Updating keyring...
-When = PostTransaction
-Exec = /usr/bin/pacman-key --init
-Exec = /usr/bin/pacman-key --populate archlinux
-_
-
 # ref: https://gitlab.archlinux.org/archlinux/archiso/-/blob/754caf0ca21476d52d8557058f665b9078982877/configs/releng/airootfs/etc/pacman.d/hooks/zzzz99-remove-custom-hooks-from-airootfs.hook
 teew etc/pacman.d/hooks/zzzz99-remove-custom-hooks-from-airootfs.hook << '_'
 # remove from airootfs!